+++ Source/SwiftLintFramework/Rules/Style/TypeContentsOrderRule.swift
@@ -1,6 +1,6 @@
 import SourceKittenFramework
 
-public struct TypeContentsOrderRule: ConfigurationProviderRule, OptInRule, ManuallyTestedExamplesRule {
+public struct TypeContentsOrderRule: ConfigurationProviderRule, OptInRule {
     private typealias TypeContentOffset = (typeContent: TypeContent, offset: ByteCount)
 
     public var configuration = TypeContentsOrderConfiguration()